refactor: remove old Main component
