+++ my-app/src/helpers/controllers/getCharacters.ts
@@ -1,5 +1,5 @@
 import axios, { AxiosError } from 'axios';
-import { ICharacter, IDocs } from 'components/Main/Main';
+import { ICharacter, IDocs } from 'components/Main/MainHooks';
 import { Mode } from 'helpers/constants/mode';
 import { routes } from 'helpers/constants/routes';
